// src/types/SEO/SEO.ts
export interface SEOData {
  Data: {
    keyword: string;
    copyright: string;
    description: string;
    author: string;
    og_title: string;
    og_description: string;
    og_image: string;
    og_type: string;
    twitter_title: string;
    twitter_description: string;
    twitter_image: string;
  }
}